+++ src/main/kotlin/uk/co/nickthecoder/tickle/action/CenterView.kt
@@ -1,11 +1,11 @@
 package uk.co.nickthecoder.tickle.action
 
-import org.joml.Vector2f
+import org.joml.Vector2d
 import uk.co.nickthecoder.tickle.stage.StageView
 
 class CenterView(
         val stageView: StageView,
-        val position: Vector2f)
+        val position: Vector2d)
 
     : Action {
