+++ src/main/kotlin/uk/co/nickthecoder/tickle/action/CenterViewBetween.kt
@@ -1,6 +1,6 @@
 package uk.co.nickthecoder.tickle.action
 
-import org.joml.Vector2f
+import org.joml.Vector2d
 import uk.co.nickthecoder.tickle.action.animation.AnimationAction
 import uk.co.nickthecoder.tickle.action.animation.Ease
 import uk.co.nickthecoder.tickle.action.animation.Eases
@@ -8,9 +8,9 @@ import uk.co.nickthecoder.tickle.stage.StageView
 
 class CenterViewBetween(
         val stageView: StageView,
-        val positionA: Vector2f,
-        val positionB: Vector2f,
-        seconds: Float = 0.5f,
+        val positionA: Vector2d,
+        val positionB: Vector2d,
+        seconds: Double = 0.5,
         ease: Ease = Eases.easeInOut)
 
     : AnimationAction(seconds, ease) {
@@ -18,7 +18,7 @@ class CenterViewBetween(
     override fun storeInitialValue() {
     }
 
-    override fun update(t: Float) {
+    override fun update(t: Double) {
         stageView.centerX = lerp(positionA.x, positionB.x, t)
         stageView.centerY = lerp(positionA.y, positionB.y, t)
     }
