+++ src/main/kotlin/uk/co/nickthecoder/tickle/action/animation/Ease.kt
@@ -2,11 +2,11 @@ package uk.co.nickthecoder.tickle.action.animation
 
 
 interface Ease {
-    fun ease(t: Float): Float
+    fun ease(t: Double): Double
 }
 
 class LinearEase : Ease {
-    override fun ease(t: Float) = t
+    override fun ease(t: Double) = t
 
     companion object {
         val instance = LinearEase()
