+++ src/main/kotlin/uk/co/nickthecoder/tickle/action/animation/Grow.kt
@@ -4,20 +4,20 @@ import uk.co.nickthecoder.tickle.Actor
 
 class Grow(
         val actor: Actor,
-        seconds: Float,
-        val finalScale: Float,
+        seconds: Double,
+        val finalScale: Double,
         ease: Ease = LinearEase.instance)
 
     : AnimationAction(seconds, ease) {
 
-    private var initialScale: Float = 0f
+    private var initialScale: Double = 0.0
 
     override fun storeInitialValue() {
         initialScale = actor.scale
     }
 
 
-    override fun update(t: Float) {
+    override fun update(t: Double) {
         actor.scale = lerp(initialScale, finalScale, t)
     }
